+++ agrest/src/test/java/io/agrest/it/POST_NaturalIdIT.java
@@ -42,7 +42,7 @@ public class POST_NaturalIdIT extends JerseyAndDerbyCase {
                 .replaceId("RID")
                 .bodyEquals(1, "{\"id\":\"John\",\"name\":\"John\"}");
 
-        e20().matcher().eq("name", "John").assertOneMatch();
+        e20().matcher().eq("name_col", "John").assertOneMatch();
 
         Response r2 = target("/single-id")
                 .queryParam("exclude", "age", "description")
@@ -51,7 +51,8 @@ public class POST_NaturalIdIT extends JerseyAndDerbyCase {
 
         onResponse(r2)
                 .statusEquals(Response.Status.BAD_REQUEST)
-                .bodyEquals("{\"success\":false,\"message\":\"Can't create 'E20' with id {name:John} - already exists\"}");
+                // TODO: DB columns exposed in the message
+                .bodyEquals("{\"success\":false,\"message\":\"Can't create 'E20' with id {name_col:John} - already exists\"}");
     }
 
     @Test