+++ agrest/src/test/java/io/agrest/it/fixture/cayenne/E27Nopk.java
@@ -1,0 +1,9 @@
+package io.agrest.it.fixture.cayenne;
+
+import io.agrest.it.fixture.cayenne.auto._E27Nopk;
+
+public class E27Nopk extends _E27Nopk {
+
+    private static final long serialVersionUID = 1L; 
+
+}