+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 Release 3.3
 
 * #414 Metadata encoding related memory leak in EncoderService
+* #415 Exposed id with name different from DB column: ExpressionException: Can't resolve path component
 
 Release 3.2